1. ShadeRec structure 2. IMaterial abstract class

+++ src/math/Mat2.h
@@ -5,19 +5,19 @@
 class Mat2 {
 public:
 	Mat2();
-	Mat2(double col0_0, double col0_1, double col1_0, double col1_1);
-	Mat2(double diag);
+	Mat2(float col0_0, float col0_1, float col1_0, float col1_1);
+	Mat2(float diag);
 	Mat2(const Vec2& a_col0, const Vec2& a_col1);
 
 	Vec2& operator[](unsigned int index);
 	Vec2 operator[](unsigned int index) const;
 
-	Mat2 operator*(double t) const;
-	const Mat2& operator*=(double t);
+	Mat2 operator*(float t) const;
+	const Mat2& operator*=(float t);
 	const Mat2& operator*=(const Mat2& m);
 
-	Mat2 operator/(double t) const;
-	const Mat2& operator/=(double t);
+	Mat2 operator/(float t) const;
+	const Mat2& operator/=(float t);
 
 	Vec2 operator*(const Vec2& u) const;
 	Mat2 operator*(const Mat2& m) const;
@@ -28,7 +28,7 @@ private:
 
 //==================================================
 
-inline Mat2 operator*(const double t, const Mat2 &m) {
+inline Mat2 operator*(const float t, const Mat2 &m) {
 	return Mat2(m[0].x*t, m[0].y*t, m[1].x*t, m[1].y*t);
 }
 
@@ -39,12 +39,12 @@ inline Mat2::Mat2() {
 	columns[1] = Vec2(0);
 }
 
-inline Mat2::Mat2(double col0_0, double col0_1, double col1_0, double col1_1) {
+inline Mat2::Mat2(float col0_0, float col0_1, float col1_0, float col1_1) {
 	columns[0] = Vec2(col0_0, col0_1);
 	columns[1] = Vec2(col1_0, col1_1);
 }
 
-inline Mat2::Mat2(double diag) {
+inline Mat2::Mat2(float diag) {
 	columns[0] = Vec2(diag, 0);
 	columns[1] = Vec2(0, diag);
 }
@@ -63,21 +63,21 @@ inline Vec2 Mat2::operator[](unsigned int index) const {
 	return columns[index];
 }
 
-inline Mat2 Mat2::operator*(double t) const {
+inline Mat2 Mat2::operator*(float t) const {
 	return Mat2(columns[0].x*t, columns[0].y*t, columns[1].x*t, columns[1].y*t);
 }
 
-inline const Mat2& Mat2::operator*=(double t) {
+inline const Mat2& Mat2::operator*=(float t) {
 	columns[0] *= t;
 	columns[1] *= t;
 	return *this;
 }
 
-inline Mat2 Mat2::operator/(double t) const {
+inline Mat2 Mat2::operator/(float t) const {
 	return Mat2(columns[0].x/t, columns[0].y/t, columns[1].x/t, columns[1].y/t);
 }
 
-inline const Mat2& Mat2::operator/=(double t) {
+inline const Mat2& Mat2::operator/=(float t) {
 	columns[0] /= t;
 	columns[1] /= t;
 	return *this;
