+++ src/math/Vec2.h
@@ -4,51 +4,51 @@
 
 class Vec2 {
 public:
-	struct { double x, y; };
+	struct { float x, y; };
 
-	Vec2(double a_x, double a_y);
-	Vec2(double a);
+	Vec2(float a_x, float a_y);
+	Vec2(float a);
 	Vec2();
 
 	const Vec2& operator+() const;
 	const Vec2 operator-() const;
 	Vec2 operator+(const Vec2 &v) const;
 	Vec2 operator-(const Vec2 &v) const;
-	Vec2 operator+(const double t) const;
-	Vec2 operator-(const double t) const;
+	Vec2 operator+(const float t) const;
+	Vec2 operator-(const float t) const;
 	const Vec2& operator=(const Vec2 &v);
-	const Vec2& operator*=(const double t);
-	const Vec2& operator/=(const double t);
+	const Vec2& operator*=(const float t);
+	const Vec2& operator/=(const float t);
 	const Vec2& operator+=(const Vec2& v);
 	const Vec2& operator-=(const Vec2& v);
-	Vec2 operator*(const double t) const;
-	Vec2 operator/(const double t) const;
+	Vec2 operator*(const float t) const;
+	Vec2 operator/(const float t) const;
 	bool operator==(const Vec2& u) const;
 
 };
 
 //==================================================
-inline Vec2 operator*(const double t, const Vec2 &u) {
+inline Vec2 operator*(const float t, const Vec2 &u) {
 	return Vec2(u.x*t, u.y*t);
 }
 
-inline Vec2 operator/(const double t, const Vec2 &u) {
+inline Vec2 operator/(const float t, const Vec2 &u) {
 	return Vec2(u.x/t, u.y/t);
 }
 
-inline Vec2 operator+(const double t, const Vec2 &u) {
+inline Vec2 operator+(const float t, const Vec2 &u) {
 	return Vec2(u.x+t, u.y+t);
 }
 
-inline Vec2 operator-(const double t, const Vec2 &u) {
+inline Vec2 operator-(const float t, const Vec2 &u) {
 	return Vec2(u.x-t, u.y-t);
 }
 
-inline double dot(const Vec2& u, const Vec2& v) {
+inline float dot(const Vec2& u, const Vec2& v) {
 	return u.x*v.x + u.y*v.y;
 }
 
-inline double length(const Vec2& u ) {
+inline float length(const Vec2& u ) {
 	return sqrt(u.x*u.x + u.y*u.y);
 }
 
@@ -62,11 +62,11 @@ inline Vec2 reflect(const Vec2& I, const Vec2& N) {
 
 //==================================================
 
-inline Vec2::Vec2(double a_x, double a_y) {
+inline Vec2::Vec2(float a_x, float a_y) {
 	x = a_x; y = a_y;
 }
 
-inline Vec2::Vec2(double a) {
+inline Vec2::Vec2(float a) {
 	x = y = a;
 }
 
@@ -90,11 +90,11 @@ inline Vec2 Vec2::operator-(const Vec2 &v) const {
 	return Vec2(x-v.x, y-v.y);
 }
 
-inline Vec2 Vec2::operator+(const double t) const {
+inline Vec2 Vec2::operator+(const float t) const {
 	return Vec2(x+t, y+t);
 }
 
-inline Vec2 Vec2::operator-(const double t) const {
+inline Vec2 Vec2::operator-(const float t) const {
 	return Vec2(x-t, y-t);
 }
 
@@ -103,12 +103,12 @@ inline const Vec2& Vec2::operator=(const Vec2 &v) {
 	return *this;
 }
 
-inline const Vec2& Vec2::operator*=(const double t) {
+inline const Vec2& Vec2::operator*=(const float t) {
 	x *= t; y *= t;
 	return *this;
 }
 
-inline const Vec2& Vec2::operator/=(const double t) {
+inline const Vec2& Vec2::operator/=(const float t) {
 	x /= t; y /= t;
 	return *this;
 }
@@ -123,11 +123,11 @@ inline const Vec2& Vec2::operator-=(const Vec2& v) {
 	return *this;
 }
 
-inline Vec2 Vec2::operator*(const double t) const {
+inline Vec2 Vec2::operator*(const float t) const {
 	return Vec2(this->x*t, this->y*t);
 }
 
-inline Vec2 Vec2::operator/(const double t) const {
+inline Vec2 Vec2::operator/(const float t) const {
 	return Vec2(this->x/t, this->y/t);
 }
