+++ src/math/Vec4.h
@@ -8,61 +8,61 @@
 class Vec4 {
 public:
 	union {
-		struct { double x, y, z, w; };
-		struct { double r, g, b, a; };
+		struct { float x, y, z, w; };
+		struct { float r, g, b, a; };
 	};
 
-	Vec4(double a_x, double a_y, double a_z, double a_w);
-	Vec4(double a);
+	Vec4(float a_x, float a_y, float a_z, float a_w);
+	Vec4(float a);
 	Vec4();
 
-	Vec4(const Vec2& xy, double a_z, double a_w);
-	Vec4(double a_x, const Vec2& yz, double a_w);
-	Vec4(double a_x, double a_y, const Vec2& zw);
+	Vec4(const Vec2& xy, float a_z, float a_w);
+	Vec4(float a_x, const Vec2& yz, float a_w);
+	Vec4(float a_x, float a_y, const Vec2& zw);
 	Vec4(const Vec2& xy, const Vec2& zw);
 
-	Vec4(const Vec3& xyz, double a_w);
-	Vec4(double a_x, const Vec3& yzw);
+	Vec4(const Vec3& xyz, float a_w);
+	Vec4(float a_x, const Vec3& yzw);
 
 	const Vec4& operator+() const;
 	const Vec4 operator-() const;
 	Vec4 operator+(const Vec4 &v) const;
 	Vec4 operator-(const Vec4 &v) const;
-	Vec4 operator+(const double t) const;
-	Vec4 operator-(const double t) const;
+	Vec4 operator+(const float t) const;
+	Vec4 operator-(const float t) const;
 	const Vec4& operator=(const Vec4 &v);
-	const Vec4& operator*=(const double t);
-	const Vec4& operator/=(const double t);
+	const Vec4& operator*=(const float t);
+	const Vec4& operator/=(const float t);
 	const Vec4& operator+=(const Vec4& v);
 	const Vec4& operator-=(const Vec4& v);
-	Vec4 operator*(const double t) const;
-	Vec4 operator/(const double t) const;
+	Vec4 operator*(const float t) const;
+	Vec4 operator/(const float t) const;
 	bool operator==(const Vec4& u) const;
 
 };
 
 //==================================================
-inline Vec4 operator*(const double t, const Vec4 &u) {
+inline Vec4 operator*(const float t, const Vec4 &u) {
 	return Vec4(u.x*t, u.y*t, u.z*t, u.w*t);
 }
 
-inline Vec4 operator/(const double t, const Vec4 &u) {
+inline Vec4 operator/(const float t, const Vec4 &u) {
 	return Vec4(u.x/t, u.y/t, u.z/t, u.w/t);
 }
 
-inline Vec4 operator+(const double t, const Vec4 &u) {
+inline Vec4 operator+(const float t, const Vec4 &u) {
 	return Vec4(u.x+t, u.y+t, u.z+t, u.w+t);
 }
 
-inline Vec4 operator-(const double t, const Vec4 &u) {
+inline Vec4 operator-(const float t, const Vec4 &u) {
 	return Vec4(u.x-t, u.y-t, u.z-t, u.w-t);
 }
 
-inline double dot(const Vec4& u, const Vec4& v) {
+inline float dot(const Vec4& u, const Vec4& v) {
 	return u.x*v.x + u.y*v.y + u.z*v.z + u.w*v.w;
 }
 
-inline double length(const Vec4& u ) {
+inline float length(const Vec4& u ) {
 	return sqrt(u.x*u.x + u.y*u.y + u.z*u.z + u.w*u.w);
 }
 
@@ -76,11 +76,11 @@ inline Vec4 reflect(const Vec4& I, const Vec4& N) {
 
 //==================================================
 
-inline Vec4::Vec4(double a_x, double a_y, double a_z, double a_w) {
+inline Vec4::Vec4(float a_x, float a_y, float a_z, float a_w) {
 	x = a_x; y = a_y; z = a_z; w = a_w;
 }
 
-inline Vec4::Vec4(double a) {
+inline Vec4::Vec4(float a) {
 	x = y = z = w = a;
 }
 
@@ -88,21 +88,21 @@ inline Vec4::Vec4() {
 	x = y = z = w = 0;
 }
 
-inline Vec4::Vec4(const Vec2& xy, double a_z, double a_w) {
+inline Vec4::Vec4(const Vec2& xy, float a_z, float a_w) {
 	x = xy.x;
 	y = xy.y;
 	z = a_z;
 	w = a_w;
 }
 
-inline Vec4::Vec4(double a_x, const Vec2& yz, double a_w) {
+inline Vec4::Vec4(float a_x, const Vec2& yz, float a_w) {
 	x = a_x;
 	y = yz.x;
 	z = yz.y;
 	w = a_w;
 }
 
-inline Vec4::Vec4(double a_x, double a_y, const Vec2& zw) {
+inline Vec4::Vec4(float a_x, float a_y, const Vec2& zw) {
 	x = a_x;
 	y = a_y;
 	z = zw.x;
@@ -116,14 +116,14 @@ inline Vec4::Vec4(const Vec2& xy, const Vec2& zw) {
 	w = zw.y;
 }
 
-inline Vec4::Vec4(const Vec3& xyz, double a_w) {
+inline Vec4::Vec4(const Vec3& xyz, float a_w) {
 	x = xyz.x;
 	y = xyz.y;
 	z = xyz.z;
 	w = a_w;
 }
 
-inline Vec4::Vec4(double a_x, const Vec3& yzw) {
+inline Vec4::Vec4(float a_x, const Vec3& yzw) {
 	x = a_x;
 	y = yzw.x;
 	z = yzw.y;
@@ -146,11 +146,11 @@ inline Vec4 Vec4::operator-(const Vec4 &v) const {
 	return Vec4(x-v.x, y-v.y, z-v.z, w-v.w);
 }
 
-inline Vec4 Vec4::operator+(const double t) const {
+inline Vec4 Vec4::operator+(const float t) const {
 	return Vec4(x+t, y+t, z+t, w+t);
 }
 
-inline Vec4 Vec4::operator-(const double t) const {
+inline Vec4 Vec4::operator-(const float t) const {
 	return Vec4(x-t, y-t, z-t, w-t);
 }
 
@@ -159,12 +159,12 @@ inline const Vec4& Vec4::operator=(const Vec4 &v) {
 	return *this;
 }
 
-inline const Vec4& Vec4::operator*=(const double t) {
+inline const Vec4& Vec4::operator*=(const float t) {
 	x *= t; y *= t; z *= t; w *= t;
 	return *this;
 }
 
-inline const Vec4& Vec4::operator/=(const double t) {
+inline const Vec4& Vec4::operator/=(const float t) {
 	x /= t; y /= t; z /= t; w /= t;
 	return *this;
 }
@@ -179,11 +179,11 @@ inline const Vec4& Vec4::operator-=(const Vec4& v) {
 	return *this;
 }
 
-inline Vec4 Vec4::operator*(const double t) const {
+inline Vec4 Vec4::operator*(const float t) const {
 	return Vec4(this->x*t, this->y*t, this->z*t, this->w*t);
 }
 
-inline Vec4 Vec4::operator/(const double t) const {
+inline Vec4 Vec4::operator/(const float t) const {
 	return Vec4(this->x/t, this->y/t, this->z/t, this->w/t);
 }
